requirements.txt
telethon
python-dotenv
pandas
psycopg2-binary
sqlalchemy
tqdm
dbt-postgres